refactor:postmanagerPage에 있던 함수들 컴포넌트안에서 호출되게 분리
