+++ src/features/comment/ui/CommentAddDialog.tsx
@@ -4,6 +4,27 @@ import { Textarea } from "@/shared/ui/TextArea"
 
 export function CommentAddDialog() {
   //showAddCommentDialog,setShowAddCommentDialog,newComment,setNewComment,addComment필요
+
+  // 댓글 추가
+  const addComment = async () => {
+    try {
+      const response = await fetch("/api/comments/add", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(newComment),
+      })
+      const data = await response.json()
+      setComments((prev) => ({
+        ...prev,
+        [data.postId]: [...(prev[data.postId] || []), data],
+      }))
+      setShowAddCommentDialog(false)
+      setNewComment({ body: "", postId: null, userId: 1 })
+    } catch (error) {
+      console.error("댓글 추가 오류:", error)
+    }
+  }
+
   return (
     <Dialog open={showAddCommentDialog} onOpenChange={setShowAddCommentDialog}>
       <DialogContent>
