+++ src/features/comment/ui/CommentEditDialog.tsx
@@ -7,6 +7,26 @@ import { Textarea } from "@/shared/ui/TextArea"
 }
 export function CommentEditDialog() {
   // showEditCommentDialog,setShowEditCommentDialog,selectedComment,setSelectedComment,updateComment값 필요
+
+  // 댓글 업데이트
+  const updateComment = async () => {
+    try {
+      const response = await fetch(`/api/comments/${selectedComment.id}`, {
+        method: "PUT",
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({ body: selectedComment.body }),
+      })
+      const data = await response.json()
+      setComments((prev) => ({
+        ...prev,
+        [data.postId]: prev[data.postId].map((comment) => (comment.id === data.id ? data : comment)),
+      }))
+      setShowEditCommentDialog(false)
+    } catch (error) {
+      console.error("댓글 업데이트 오류:", error)
+    }
+  }
+
   return (
     <Dialog open={showEditCommentDialog} onOpenChange={setShowEditCommentDialog}>
       <DialogContent>
