+++ src/features/post/ui/PostSearchFilter.tsx
@@ -10,6 +10,24 @@ export function PostSearchFilter() {
 
   const updateURL = usePostUpdateURL()
 
+  // 게시물 검색
+  const searchPosts = async () => {
+    if (!searchQuery) {
+      fetchPosts()
+      return
+    }
+    setLoading(true)
+    try {
+      const response = await fetch(`/api/posts/search?q=${searchQuery}`)
+      const data = await response.json()
+      setPosts(data.posts)
+      setTotal(data.total)
+    } catch (error) {
+      console.error("게시물 검색 오류:", error)
+    }
+    setLoading(false)
+  }
+
   return (
     <div className="flex gap-4">
       <div className="flex-1">
